requirements.txt
flask
openai==0.28
requests
sendgrid
gunicorn
python-dotenv
